requirements.txt
python-dateutil
pandas
matplotlib
seaborn
numpy